# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AeroHub API
  description: 'API operacional de AeroHub: aprovisionamiento de tenants, seguimiento
    de vuelos y distribucion de FIDS en tiempo real (RF-T02, OpenAPI 3.1).'
  contact:
    name: AeroHub Platform Team
  version: 0.1.0
servers:
- url: http://localhost:8000
  description: Desarrollo local
paths:
...
  /support/tickets:
    post:
      tags:
      - soporte
      summary: Crear Ticket Endpoint
      operationId: crear_ticket_endpoint_support_tickets_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearTicketRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrearTicketResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - soporte
      summary: Listar Tickets Endpoint
      operationId: listar_tickets_endpoint_support_tickets_get
      parameters:
      - name: estado
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Estado
      - name: severidad
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Severidad
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketResponse'
                title: Response Listar Tickets Endpoint Support Tickets Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /support/tickets/{ticket_id}:
    get:
      tags:
      - soporte
      summary: Obtener Ticket Endpoint
      operationId: obtener_ticket_endpoint_support_tickets__ticket_id__get
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: integer
          title: Ticket Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetalleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /support/tickets/{ticket_id}/transiciones:
    get:
      tags:
      - soporte
      summary: Listar Transiciones Ticket Endpoint
      description: 'Linea de tiempo de cambios de estado del ticket (Fase 3 de

        docs/diseno/PLAN_CORRECCION_MODULOS.md, item 8): completa el hilo de

        mensajes con quien cambio el estado, cuando y a que -- antes de este

        endpoint ese dato ya se registraba (compliance.log_auditoria) pero

        nunca se exponia.'
      operationId: listar_transiciones_ticket_endpoint_support_tickets__ticket_id__transiciones_get
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: integer
          title: Ticket Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransicionResponse'
                title: Response Listar Transiciones Ticket Endpoint Support Tickets  Ticket
                  Id  Transiciones Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /support/tickets/{ticket_id}/mensajes:
    post:
      tags:
      - soporte
      summary: Responder Ticket Endpoint
      operationId: responder_ticket_endpoint_support_tickets__ticket_id__mensajes_post
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: integer
          title: Ticket Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponderTicketRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponderTicketResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /support/tickets/{ticket_id}/estado:
    patch:
      tags:
      - soporte
      summary: Cambiar Estado Ticket Endpoint
      operationId: cambiar_estado_ticket_endpoint_support_tickets__ticket_id__estado_patch
      parameters:
      - name: ticket_id
        in: path
        required: true
        schema:
          type: integer
          title: Ticket Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CambiarEstadoTicketRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: boolean
                title: Response Cambiar Estado Ticket Endpoint Support Tickets  Ticket
                  Id  Estado Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /support/kb/articulos:
    post:
      tags:
      - soporte
      summary: Publicar Articulo Endpoint
      operationId: publicar_articulo_endpoint_support_kb_articulos_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicarArticuloRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicarArticuloResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - soporte
      summary: Buscar Articulos Endpoint
      operationId: buscar_articulos_endpoint_support_kb_articulos_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: etiqueta
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Etiqueta
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArticuloResponse'
                title: Response Buscar Articulos Endpoint Support Kb Articulos Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /support/kb/articulos/{articulo_id}:
    get:
      tags:
      - soporte
      summary: Obtener Articulo Endpoint
      operationId: obtener_articulo_endpoint_support_kb_articulos__articulo_id__get
      parameters:
      - name: articulo_id
        in: path
        required: true
        schema:
          type: integer
          title: Articulo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticuloResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /support/changelog:
    get:
      tags:
      - soporte
      summary: Listar Changelog Endpoint
      operationId: listar_changelog_endpoint_support_changelog_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ChangelogResponse'
                type: array
                title: Response Listar Changelog Endpoint Support Changelog Get
    post:
      tags:
      - soporte
      summary: Publicar Changelog Endpoint
      operationId: publicar_changelog_endpoint_support_changelog_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicarChangelogRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicarChangelogResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /support/observabilidad/uptime:
    get:
      tags:
      - soporte
      summary: Obtener Uptime Endpoint
      operationId: obtener_uptime_endpoint_support_observabilidad_uptime_get
      parameters:
      - name: servicio
        in: query
        required: true
        schema:
          enum:
          - aodb
          - fids
          type: string
          title: Servicio
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservabilidadUptimeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ArticuloResponse:
      properties:
        id:
          type: string
          title: Id
        titulo:
          type: string
          title: Titulo
        cuerpo:
          type: string
          title: Cuerpo
        version:
          type: integer
          title: Version
        publicado_en:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Publicado En
        etiquetas:
          items:
            type: string
          type: array
          title: Etiquetas
      type: object
      required:
      - id
      - titulo
      - cuerpo
      - version
      - publicado_en
      - etiquetas
      title: ArticuloResponse
...
    CambiarEstadoTicketRequest:
      properties:
        estado:
          type: string
          enum:
          - en_progreso
          - esperando_cliente
          - resuelto
          - cerrado
          title: Estado
      type: object
      required:
      - estado
      title: CambiarEstadoTicketRequest
...
    ChangelogResponse:
      properties:
        id:
          type: string
          title: Id
        version_producto:
          type: string
          title: Version Producto
        resumen:
          type: string
          title: Resumen
        publicado_en:
          type: string
          format: date-time
          title: Publicado En
        items:
          items:
            $ref: '#/components/schemas/ItemChangelogResponse'
          type: array
          title: Items
      type: object
      required:
      - id
      - version_producto
      - resumen
      - publicado_en
      - items
      title: ChangelogResponse
...
    CrearTicketRequest:
      properties:
        categoria_id:
          type: string
          title: Categoria Id
        severidad:
          type: string
          enum:
          - baja
          - media
          - alta
          - critica
          title: Severidad
        asunto:
          type: string
          title: Asunto
        cuerpo_inicial:
          type: string
          title: Cuerpo Inicial
      type: object
      required:
      - categoria_id
      - severidad
      - asunto
      - cuerpo_inicial
      title: CrearTicketRequest
    CrearTicketResponse:
      properties:
        ticket_id:
          type: string
          title: Ticket Id
        sla_objetivo_min:
          type: integer
          title: Sla Objetivo Min
      type: object
      required:
      - ticket_id
      - sla_objetivo_min
      title: CrearTicketResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ItemChangelogRequest:
      properties:
        modulo_codigo:
          type: string
          title: Modulo Codigo
        tipo_cambio:
          type: string
          enum:
          - nuevo
          - mejora
          - correccion
          - obsolescencia
          title: Tipo Cambio
        descripcion:
          type: string
          title: Descripcion
      type: object
      required:
      - modulo_codigo
      - tipo_cambio
      - descripcion
      title: ItemChangelogRequest
    ItemChangelogResponse:
      properties:
        id:
          type: string
          title: Id
        modulo_id:
          type: string
          title: Modulo Id
        tipo_cambio:
          type: string
          title: Tipo Cambio
        descripcion:
          type: string
          title: Descripcion
      type: object
      required:
      - id
      - modulo_id
      - tipo_cambio
      - descripcion
      title: ItemChangelogResponse
...
    MensajeResponse:
      properties:
        id:
          type: string
          title: Id
        autor_usuario_id:
          type: string
          title: Autor Usuario Id
        cuerpo:
          type: string
          title: Cuerpo
        enviado_en:
          type: string
          format: date-time
          title: Enviado En
        es_interno:
          type: boolean
          title: Es Interno
      type: object
      required:
      - id
      - autor_usuario_id
      - cuerpo
      - enviado_en
      - es_interno
      title: MensajeResponse
...
    ObservabilidadUptimeResponse:
      properties:
        servicio:
          type: string
          title: Servicio
        uptime_pct:
          type: number
          title: Uptime Pct
        error_budget_consumido_pct:
          type: number
          title: Error Budget Consumido Pct
      type: object
      required:
      - servicio
      - uptime_pct
      - error_budget_consumido_pct
      title: ObservabilidadUptimeResponse
...
    PublicarArticuloRequest:
      properties:
        titulo:
          type: string
          title: Titulo
        cuerpo:
          type: string
          title: Cuerpo
        etiquetas:
          items:
            type: string
          type: array
          title: Etiquetas
          default: []
      type: object
      required:
      - titulo
      - cuerpo
      title: PublicarArticuloRequest
    PublicarArticuloResponse:
      properties:
        articulo_id:
          type: string
          title: Articulo Id
        version:
          type: integer
          title: Version
      type: object
      required:
      - articulo_id
      - version
      title: PublicarArticuloResponse
    PublicarChangelogRequest:
      properties:
        version_producto:
          type: string
          title: Version Producto
        resumen:
          type: string
          title: Resumen
        items:
          items:
            $ref: '#/components/schemas/ItemChangelogRequest'
          type: array
          title: Items
      type: object
      required:
      - version_producto
      - resumen
      - items
      title: PublicarChangelogRequest
    PublicarChangelogResponse:
      properties:
        changelog_id:
          type: string
          title: Changelog Id
      type: object
      required:
      - changelog_id
      title: PublicarChangelogResponse
...
    ResponderTicketRequest:
      properties:
        cuerpo:
          type: string
          title: Cuerpo
        es_interno:
          type: boolean
          title: Es Interno
          default: false
      type: object
      required:
      - cuerpo
      title: ResponderTicketRequest
    ResponderTicketResponse:
      properties:
        mensaje_id:
          type: string
          title: Mensaje Id
      type: object
      required:
      - mensaje_id
      title: ResponderTicketResponse
...
    TicketDetalleResponse:
      properties:
        ticket:
          $ref: '#/components/schemas/TicketResponse'
        mensajes:
          items:
            $ref: '#/components/schemas/MensajeResponse'
          type: array
          title: Mensajes
      type: object
      required:
      - ticket
      - mensajes
      title: TicketDetalleResponse
    TicketResponse:
      properties:
        id:
          type: string
          title: Id
        tenant_id:
          type: string
          title: Tenant Id
        categoria_id:
          type: string
          title: Categoria Id
        creado_por_usuario_id:
          type: string
          title: Creado Por Usuario Id
        asignado_a_usuario_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Asignado A Usuario Id
        severidad:
          type: string
          title: Severidad
        estado:
          type: string
          title: Estado
        asunto:
          type: string
          title: Asunto
        creado_en:
          type: string
          format: date-time
          title: Creado En
        primera_respuesta_en:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Primera Respuesta En
        resuelto_en:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Resuelto En
        sla_objetivo_min:
          type: integer
          title: Sla Objetivo Min
      type: object
      required:
      - id
      - tenant_id
      - categoria_id
      - creado_por_usuario_id
      - asignado_a_usuario_id
      - severidad
      - estado
      - asunto
      - creado_en
      - primera_respuesta_en
      - resuelto_en
      - sla_objetivo_min
      title: TicketResponse
...
    TransicionResponse:
      properties:
        id:
          type: string
          title: Id
        usuario_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Usuario Id
        rol_codigo:
          type: string
          title: Rol Codigo
        estado_anterior:
          anyOf:
          - type: string
          - type: 'null'
          title: Estado Anterior
        estado_nuevo:
          anyOf:
          - type: string
          - type: 'null'
          title: Estado Nuevo
        ocurrido_en:
          type: string
          format: date-time
          title: Ocurrido En
      type: object
      required:
      - id
      - usuario_id
      - rol_codigo
      - estado_anterior
      - estado_nuevo
      - ocurrido_en
      title: TransicionResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...